// template/react-starter-kit/src/.generated/api/model/testWebhookRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { TestWebhookRequestDtoType } from './testWebhookRequestDtoType';

export interface TestWebhookRequestDto {
  /** 알림 채널 종류 */
  type: TestWebhookRequestDtoType;
  /** 테스트 전송할 웹훅 URL */
  webhookUrl: string;
}